+++ changelog.md
@@ -9,6 +9,7 @@
 ## 🐣beta-4 WIP
 
 - ✨ Improved quite a few existing nodes.
+- ✨ Node pins now have a color that reflects the kind of node that you can plug into them. (e.g. a pin that wants a Shape 3D will be yellow, just like the Shape 3D nodes.)
 - ✨ Smarter automatic main node selection.
 - ✨ All nodes can now be viewed by themselves, without requiring a "renderer" node (which used to be the case for Curves and 3D Shapes).
 - ✨ Added buttons on the view to freeze / enable the 2D and 3D cameras.